Familias profesionales y ciclos test correctos

+++ app/Http/Controllers/API/CicloFormativoController.php
@@ -7,17 +7,18 @@ use App\Models\CicloFormativo;
 use Illuminate\Http\Request;
 use App\Http\Resources\CicloFormativoResource;
 use App\Models\FamiliaProfesional;
+use Illuminate\Support\Facades\Auth;
 
 class CicloFormativoController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index(Request $request, FamiliaProfesional $familiaProfesional)
+    public function index(Request $request, $FamiliaProfesionalId)
     {
-        $query = CicloFormativo::where('familia_profesional_id', $familiaProfesional->id);
+        $query = CicloFormativo::where('familia_profesional_id', $FamiliaProfesionalId);
         if ($query) {
-            $query->where('nombre', 'like', '%' . $request->q . '%');
+            $query->where('nombre', 'like', '%' . $request->search . '%');
         }
 
         return CicloFormativoResource::collection(
@@ -29,11 +30,21 @@ class CicloFormativoController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(Request $request, $FamiliaProfesionalId)
     {
-        $cicloFormativo = json_decode($request->getContent(), true);
-
-        $cicloFormativo = CicloFormativo::create($cicloFormativo);
+        if(Auth::user()->email != env('ADMIN_EMAIL')) {
+            return response()->json([
+                'message' => 'Unauthorized'
+            ], 403);
+        }
+        $validate_data = $request->validate([
+            'nombre' => 'required|string|max:255',
+            'codigo' => 'required|string|max:50|unique:ciclos_formativos,codigo',
+            'grado' => 'required|string|max:50|in:' . implode(',', CicloFormativo::GRADOS),
+            'descripcion' => 'nullable|string|max:255',
+        ]);
+        $validate_data['familia_profesional_id'] = $FamiliaProfesionalId;
+        $cicloFormativo = CicloFormativo::create($validate_data);
 
         return new CicloFormativoResource($cicloFormativo);
     }
@@ -48,10 +59,21 @@ class CicloFormativoController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, FamiliaProfesional $familiaProfesional, CicloFormativo $cicloFormativo)
+    public function update(Request $request, $FamiliaProfesionalId , CicloFormativo $cicloFormativo)
     {
-        $cicloFormativoData = json_decode($request->getContent(), true);
-        $cicloFormativo->update($cicloFormativoData);
+        if(Auth::user()->email != env('ADMIN_EMAIL')) {
+            return response()->json([
+                'message' => 'Unauthorized'
+            ], 403);
+        }
+        $validate_data = $request->validate([
+            'nombre' => 'required|string|max:255',
+            'codigo' => 'required|string|max:50|unique:ciclos_formativos,codigo,',
+            'grado' => 'required|string|max:50|in:' . implode(',', CicloFormativo::GRADOS),
+            'descripcion' => 'nullable|string|max:255',
+        ]);
+        $validate_data['familia_profesional_id'] = $FamiliaProfesionalId;
+        $cicloFormativo->update($validate_data);
 
         return new CicloFormativoResource($cicloFormativo);
     }
@@ -61,9 +83,16 @@ class CicloFormativoController extends Controller
      */
     public function destroy(FamiliaProfesional $familiaProfesional, CicloFormativo $cicloFormativo)
     {
+        if(Auth::user()->email != env('ADMIN_EMAIL')) {
+            return response()->json([
+                'message' => 'Unauthorized'
+            ], 403);
+        }
         try {
             $cicloFormativo->delete();
-            return response()->json(null, 204);
+            return response()->json([
+                'message' => 'CicloFormativo eliminado correctamente'
+            ], 200);
         } catch (\Exception $e) {
             return response()->json([
                 'message' => 'Error: ' . $e->getMessage()
